+++ CHANGELOG.md
@@ -28,6 +28,7 @@ _This release is scheduled to be released on 2025-01-01._
 - [core] Fix loading node_helper of modules: avoid black screen, display errors and continue loading with next module (#3578)
 - [weather] changed default value for weatherEndpoint of provider openweathermap to "/onecall" (#3574)
 - [calendar] - update to resolve issues #3098 #3144 #3351 #3422 #3443 #3467 #3537 related to timezone changes
+- [calendar] - fixes #3267 (styles array), also fixes event with both exdate AND recurrence(and testcase)
 
 ## [2.29.0] - 2024-10-01